day03: seekg, pair, bitset, list::erase
operator>> for bitset took care of parsing the input.
The filtering procedure for part2 took some time to write, as I
digressed trying to find initially a nearest hamming distance
neighbour, and then a weighted distance neighbour to the average of
part1. After I realised that they both are wrong, I just implemented
the procedure from the problem verbatim.